backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
pydantic[email]==2.11.3
mongoengine==0.28.2
pymongo==4.6.0
werkzeug==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.10.1
pwdlip==1.7.4